AIproject/requirements.txt
absl-py==2.2.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.1
Flask==3.1.1
flask-cors==6.0.1
Flask-MySQLdb==2.0.0
flatbuffers==25.2.10
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mysql==0.0.3
mysqlclient==2.2.7
namex==0.0.9
narwhals==2.0.1
numpy==2.1.3
openai==1.78.1
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.4
pyarrow==21.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
qrcode==8.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.27.0
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.2
streamlit==1.48.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
